// src/types/settings.ts
export interface TaskNotesSettings {
// ...
	fieldMapping: FieldMapping;
// ...
}